Rename Quill\Format to \Formats\RichFormat
